refactor: move StreamTester2
